// apps/frontend2/src/api/models/createExperienceRequest.ts
/**
 * Generated by orval v7.11.2 🍺
 * Do not edit manually.
 * FastAPI
 * OpenAPI spec version: 0.1.0
 */
import type { CreateExperienceRequestEndDate } from './createExperienceRequestEndDate';

export interface CreateExperienceRequest {
  /** @minLength 1 */
  company: string;
  /** @minLength 1 */
  title: string;
  start_date: string;
  description: string;
  end_date?: CreateExperienceRequestEndDate;
}