// apps/frontend2/src/api/models/profileExperienceDTO.ts
/**
 * Generated by orval v7.11.2 🍺
 * Do not edit manually.
 * FastAPI
 * OpenAPI spec version: 0.1.0
 */
import type { ProfileExperienceDTOEndDate } from './profileExperienceDTOEndDate';

export interface ProfileExperienceDTO {
  id: number;
  company: string;
  title: string;
  description: string;
  start_date: string;
  end_date: ProfileExperienceDTOEndDate;
  is_current: boolean;
  created_at: string;
  updated_at: string;
}