// apps/frontend2/src/api/models/updateExperienceRequest.ts
/**
 * Generated by orval v7.11.2 🍺
 * Do not edit manually.
 * FastAPI
 * OpenAPI spec version: 0.1.0
 */
import type { UpdateExperienceRequestCompany } from './updateExperienceRequestCompany';
import type { UpdateExperienceRequestTitle } from './updateExperienceRequestTitle';
import type { UpdateExperienceRequestDescription } from './updateExperienceRequestDescription';
import type { UpdateExperienceRequestStartDate } from './updateExperienceRequestStartDate';
import type { UpdateExperienceRequestEndDate } from './updateExperienceRequestEndDate';

export interface UpdateExperienceRequest {
  company?: UpdateExperienceRequestCompany;
  title?: UpdateExperienceRequestTitle;
  description?: UpdateExperienceRequestDescription;
  start_date?: UpdateExperienceRequestStartDate;
  end_date?: UpdateExperienceRequestEndDate;
}